+++ gazebo/gui/GLWidget.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2012-2014 Open Source Robotics Foundation
+ * Copyright (C) 2012-2015 Open Source Robotics Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -286,7 +286,8 @@ void GLWidget::keyPressEvent(QKeyEvent *_event)
 
   // Trigger a model delete if the Delete key was pressed, and a model
   // is currently selected.
-  if (_event->key() == Qt::Key_Delete)
+  if (_event->key() == Qt::Key_Delete &&
+      this->selectionLevel == SelectionLevels::MODEL)
   {
     boost::mutex::scoped_lock lock(this->selectedVisMutex);
     while (!this->selectedVisuals.empty())
@@ -691,15 +692,75 @@ void GLWidget::OnMouseReleaseNormal()
   {
     rendering::VisualPtr vis =
       this->userCamera->GetVisual(this->mouseEvent.pos);
+
     if (vis)
     {
-      vis = vis->GetRootVisual();
-      this->SetSelectedVisual(vis);
-      event::Events::setSelectedEntity(vis->GetName(), "normal");
-
-      if (this->mouseEvent.button == common::MouseEvent::RIGHT)
+      rendering::VisualPtr selectVis;
+      rendering::VisualPtr linkVis = vis->GetParent();
+      if (!linkVis)
       {
-        g_modelRightMenu->Run(vis->GetName(), QCursor::pos());
+        gzerr << "Link visual not found, this should not happen." << std::endl;
+        return;
+      }
+      rendering::VisualPtr modelVis = vis->GetRootVisual();
+      if (!modelVis)
+      {
+        gzerr << "Model visual not found, this should not happen." << std::endl;
+        return;
+      }
+
+      // Flags to check if we should select a link or a model
+      bool rightButton = (this->mouseEvent.button == common::MouseEvent::RIGHT);
+      bool modelHighlighted = modelVis->GetHighlighted();
+      int linkCount = 0;
+      bool linkHighlighted = false;
+      for (unsigned int i = 0; i < modelVis->GetChildCount(); ++i)
+      {
+        // Find out if there's only one link in the model
+        uint32_t flags = modelVis->GetChild(i)->GetVisibilityFlags();
+        if ((flags != GZ_VISIBILITY_ALL) && (flags & GZ_VISIBILITY_GUI))
+        {
+          continue;
+        }
+        linkCount++;
+
+        // A link from the same model is currently selected
+        if (modelVis->GetChild(i)->GetHighlighted())
+        {
+          linkHighlighted = true;
+        }
+      }
+
+      // Select link
+      if (linkCount > 1 && !this->mouseEvent.control &&
+          ((modelHighlighted && !rightButton) || linkHighlighted))
+      {
+        selectVis = linkVis;
+      }
+      // Select model
+      else
+      {
+        // Can't select a link and a model at the same time
+        if (this->selectionLevel == SelectionLevels::LINK)
+          this->DeselectAllVisuals();
+
+        selectVis = modelVis;
+      }
+      this->SetSelectedVisual(selectVis);
+      event::Events::setSelectedEntity(selectVis->GetName(), "normal");
+
+      // Open context menu
+      if (rightButton)
+      {
+        if (selectVis == modelVis)
+        {
+          g_modelRightMenu->Run(selectVis->GetName(), QCursor::pos(),
+              ModelRightMenu::EntityTypes::MODEL);
+        }
+        else if (selectVis == linkVis)
+        {
+          // TODO: Open link right menu
+        }
       }
     }
     else
@@ -936,26 +997,23 @@ void GLWidget::OnSelectionMsg(ConstSelectionPtr &_msg)
 /////////////////////////////////////////////////
 void GLWidget::SetSelectedVisual(rendering::VisualPtr _vis)
 {
+  // deselect all if not in multi-selection mode.
+  if (!this->mouseEvent.control)
+  {
+    this->DeselectAllVisuals();
+  }
+
   boost::mutex::scoped_lock lock(this->selectedVisMutex);
 
   msgs::Selection msg;
 
-  // deselect all if not in multi-selection mode.
-  if (!this->mouseEvent.control)
-  {
-    for (unsigned int i = 0; i < this->selectedVisuals.size(); ++i)
-    {
-      this->selectedVisuals[i]->SetHighlighted(false);
-      msg.set_id(this->selectedVisuals[i]->GetId());
-      msg.set_name(this->selectedVisuals[i]->GetName());
-      msg.set_selected(false);
-      this->selectionPub->Publish(msg);
-    }
-    this->selectedVisuals.clear();
-  }
-
   if (_vis && !_vis->IsPlane())
   {
+    if (_vis == _vis->GetRootVisual())
+      this->selectionLevel = SelectionLevels::MODEL;
+    else
+      this->selectionLevel = SelectionLevels::LINK;
+
     _vis->SetHighlighted(true);
 
     // enable multi-selection if control is pressed
@@ -987,6 +1045,23 @@ void GLWidget::SetSelectedVisual(rendering::VisualPtr _vis)
   }
 
   g_alignAct->setEnabled(this->selectedVisuals.size() > 1);
+}
+
+/////////////////////////////////////////////////
+void GLWidget::DeselectAllVisuals()
+{
+  boost::mutex::scoped_lock lock(this->selectedVisMutex);
+
+  msgs::Selection msg;
+  for (unsigned int i = 0; i < this->selectedVisuals.size(); ++i)
+  {
+    this->selectedVisuals[i]->SetHighlighted(false);
+    msg.set_id(this->selectedVisuals[i]->GetId());
+    msg.set_name(this->selectedVisuals[i]->GetName());
+    msg.set_selected(false);
+    this->selectionPub->Publish(msg);
+  }
+  this->selectedVisuals.clear();
 }
 
 /////////////////////////////////////////////////
@@ -1172,6 +1247,7 @@ void GLWidget::OnRequest(ConstRequestPtr &_msg)
       {
         if ((*it)->GetName() == _msg->data())
         {
+          ModelManipulator::Instance()->Detach();
           this->selectedVisuals.erase(it);
           break;
         }
@@ -1201,11 +1277,6 @@ void GLWidget::OnModelEditor(bool _checked)
   g_arrowAct->trigger();
   event::Events::setSelectedEntity("", "normal");
 
-  boost::mutex::scoped_lock lock(this->selectedVisMutex);
   // Manually deselect, in case the editor was opened with Ctrl
-  for (unsigned int i = 0; i < this->selectedVisuals.size(); ++i)
-  {
-    this->selectedVisuals[i]->SetHighlighted(false);
-  }
-  this->selectedVisuals.clear();
+  this->DeselectAllVisuals();
 }